feat: add clean database method
