+++ apps/backend/package.json
@@ -21,7 +21,8 @@
     "test:e2e": "jest --config ./test/jest-e2e.json",
     "seed": "node dist/prisma/seed.js",
     "seed:bun": "bun dist/prisma/seed.js",
-    "migrate": "bun prisma migrate dev"
+    "migrate": "bun prisma migrate dev",
+    "clean:db": "node dist/prisma/clean.js"
   },
   "dependencies": {
     "@aws-sdk/client-s3": "^3.627.0",
@@ -31,6 +32,7 @@
     "@nestjs/core": "^10.3.10",
     "@nestjs/jwt": "^10.2.0",
     "@nestjs/platform-express": "^10.3.10",
+    "@nestjs/schedule": "^4.1.0",
     "@nestjs/swagger": "^7.4.0",
     "@nestjs/throttler": "^6.1.0",
     "@paralleldrive/cuid2": "^2.2.2",
