+++ apps/backend/prisma/schema.prisma
@@ -39,7 +39,7 @@ enum UserType {
 model Product {
   id String @id @default(uuid())
 
-  name        String
+  name        String  @unique
   description String
   slug        String  @unique
   buyPrice    Int     @map("buy_price")
@@ -138,12 +138,13 @@ model UserPurchases {
   productId String  @unique @map("product_id")
   product   Product @relation(fields: [productId], references: [id])
 
-  productLink String @map("product_link")
-  productSlug String @map("product_slug")
-  productValue Int @map("product_values")
-  productName String @map("product_name")
+  productLink  String @map("product_link")
+  productSlug  String @map("product_slug")
+  productValue Int    @map("product_values")
+  productName  String @map("product_name")
 
-  createdAt     DateTime       @default(now()) @map("created_at")
-  updatedAt     DateTime       @updatedAt @map("updated_at")
+  createdAt DateTime @default(now()) @map("created_at")
+  updatedAt DateTime @updatedAt @map("updated_at")
+
   @@map("user_purchases")
 }
