+++ apps/backend/src/prisma/clean.ts
@@ -1,0 +1,5 @@
+import { PrismaHelpers } from "src/prisma/helpers";
+
+const prismaHelpers = new PrismaHelpers();
+
+prismaHelpers.clearDatabase();
